Switch number to f64 (decimals), and replace JIT with real AOT compilation
number was backed by i64 - now f64, so declarations like
number '3.14' work. Print formats with "%.15g" (full double precision,
trailing zeros trimmed) so whole numbers still print as "1000" rather
than "1000.000000". Known side effect: literals too large for i64
(e.g. 28-digit numbers) now silently parse as an approximate f64
instead of erroring - expected, since that's what switching to a
floating-point representation means.

Replaces JIT execution entirely with ahead-of-time compilation to a
real, standalone native binary: emit an object file via LLVM's target
machine, then link it with the system 'cc' (which also gives us libc,
since we call printf). No more JIT dependency at runtime.

New CLI syntax, deliberately verbose/self-describing over terse flags:
  qmcl task:compile file:hello.qmcl [outputfilename:name]
  qmcl task:run file:hello.qmcl [outputfilename:name]
task:compile builds only (gcc/rustc convention - you run the binary
yourself); task:run builds and runs it immediately. Arguments are
parsed by matching each known key's full "key:" prefix, not by
splitting on the first colon anywhere - the latter would misparse a
Windows-style path like file:C:\Users\you\hello.qmcl at the drive
letter's colon.

+++ src/ast.rs
@@ -7,7 +7,7 @@ use crate::error::Span;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum Expr {
-    NumberLiteral(i64),
+    NumberLiteral(f64),
     /// A variable reference (the `x` in `(x)`), carrying where it was
     /// written so codegen can point at it if the variable turns out to be
     /// undeclared.
